requirements.txt
py==1.11.0
pytest==7.1.2
requests==2.27.1
allure-pytest==2.9.45
